requirements.txt
matplotlib>=3.9.0rc2
torch>=2.2.2
pillow>=10.3.0
torchvision>=0.17.2
tqdm>=4.66.2